move click stuff to core package
