+++ core/src/main/java/org/incendo/interfaces/core/click/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * Classes related to Paper click handling.
+ */
+package org.incendo.interfaces.core.click;
