+++ kotlin/src/main/kotlin/org/incendo/interfaces/kotlin/MutableInterfaceBuilder.kt
@@ -1,7 +1,7 @@
 package org.incendo.interfaces.kotlin
 
 import org.incendo.interfaces.core.pane.Pane
-import org.incendo.interfaces.paper.click.ClickHandler
+import org.incendo.interfaces.core.click.ClickHandler
 
 public interface MutableInterfaceBuilder<T : Pane> {
