+++ paper/src/main/java/org/incendo/interfaces/paper/type/ChestInterface.java
@@ -11,7 +11,7 @@ import org.incendo.interfaces.core.transform.Transform;
 import org.incendo.interfaces.core.transform.TransformContext;
 import org.incendo.interfaces.core.view.InterfaceView;
 import org.incendo.interfaces.paper.PlayerViewer;
-import org.incendo.interfaces.paper.click.ClickHandler;
+import org.incendo.interfaces.core.click.ClickHandler;
 import org.incendo.interfaces.paper.pane.ChestPane;
 import org.incendo.interfaces.paper.view.ChestView;
 import org.incendo.interfaces.paper.view.PlayerView;
